Add: http error test for test_get_star_wars_data
